Extract loss functions and wandb logging.
